backend/requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
sqlalchemy
apscheduler
python-multipart
python-dotenv
